requirements.txt
beautifulsoup4==4.8.2
requests==2.22.0
matplotlib==3.1.2
pillow==7.0.0
